% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellbe_data_connections.R
\name{looker_pull}
\alias{looker_pull}
\title{This function allows you to easily pull clean looker data from a look.}
\usage{
looker_pull(look_id, file_name = "looker_pull", re_run = FALSE,
  path = "~/wellbe-analytics/LookRApiCredentials.R")
}
\arguments{
\item{look_id}{Look ID to pull.}

\item{file_name}{Name for pull to avoid rerunning. Defaults to looker_pull.}

\item{re_run}{If TRUE, overwrites any existing data. Defauts to FALSE.}
}
\description{
This function allows you to easily pull clean looker data from a look.
}
\examples{
looker_pull()
}
\keyword{database}
