% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellbe_data_connections.R
\name{read_replica_query}
\alias{read_replica_query}
\title{This function allows you to run SQL queries from the read replica of the Wellbe Database}
\usage{
read_replica_query(sql, path = "~/wellbe-analytics/dbicred.r")
}
\arguments{
\item{path}{Path to where DB credentials are. Defauts to where credentials are saved on Andrew's work computer.}

\item{SQL}{SQL query to run.}
}
\description{
This function allows you to run SQL queries from the read replica of the Wellbe Database
}
\examples{
read_replica_query()
}
\keyword{database}
